+++ KeireCore/premake5.lua
@@ -28,8 +28,7 @@ project(ProjectConfig.CORE_TARGET)
 
     filter "system:windows"
         prebuildcommands {
-            "if exist Scripts\\Windows\\build-info.ps1 (powershell -NoProfile -ExecutionPolicy Bypass -File Scripts\\Windows\\build-info.ps1) else (powershell -NoProfile -ExecutionPolicy Bypass -File ..\\Scripts\\Windows\\build-info.ps1)",
-            "if exist Scripts\\Windows\\builtin-shaders.ps1 (powershell -NoProfile -ExecutionPolicy Bypass -File Scripts\\Windows\\builtin-shaders.ps1) else (powershell -NoProfile -ExecutionPolicy Bypass -File ..\\Scripts\\Windows\\builtin-shaders.ps1)"
+            "if exist Scripts\\Windows\\build-info.ps1 (powershell -NoProfile -ExecutionPolicy Bypass -File Scripts\\Windows\\build-info.ps1 && powershell -NoProfile -ExecutionPolicy Bypass -File Scripts\\Windows\\builtin-shaders.ps1) else (powershell -NoProfile -ExecutionPolicy Bypass -File ..\\Scripts\\Windows\\build-info.ps1 && powershell -NoProfile -ExecutionPolicy Bypass -File ..\\Scripts\\Windows\\builtin-shaders.ps1)"
         }
 
     filter { "system:linux or macosx" }